+++ src/control_plane/migration/rabbitmq/mod.rs
@@ -1,0 +1,5 @@
+mod backup_rabbitmq_vhost;
+mod rabbitmq_backup_options;
+
+pub(crate) use backup_rabbitmq_vhost::backup_rabbitmq_vhost;
+pub(crate) use rabbitmq_backup_options::RabbitMqBackupOptions;